requirements.txt
pydicom>=2.3.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scipy>=1.7.0
pillow>=8.0.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
pandas>=1.3.0